+++ order/src/main/java/com/emotionalcart/order/domain/order/entity/Money.java
@@ -1,4 +1,4 @@
-package com.emotionalcart.order.domain.entity;
+package com.emotionalcart.order.domain.order.entity;
 
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
