+++ order/src/main/java/com/emotionalcart/order/domain/order/entity/OrderAddress.java
@@ -1,4 +1,4 @@
-package com.emotionalcart.order.domain.entity;
+package com.emotionalcart.order.domain.order.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
